linux-bradfa: Fix defconfig file naming
Treat the kernel config in the layer as a config fragment, even though
it can be a full config, so that the normal config handling functions do
the right things.  It does not appear that you can have a defconfig
which is not either already in your source tree or which is not
explicitly named exactly "defconfig".

Signed-off-by: Andrew Bradford <andrew@bradfordembedded.com>